refactror: move to ts and fix configs
