+++ nodemon.json
@@ -1,0 +1,5 @@
+{
+  "execMap": {
+    "ts": "ts-node"
+  }
+}
