+++ package.json
@@ -2,13 +2,12 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "src/index.js",
+    "main": "src/index.ts",
     "type": "module",
     "scripts": {
         "build": "tsc",
-        "start": "node ./src/index.js",
-        "start:dev": "nodemon ./src/index.js",
-        "test": "echo \"Error: no test specified\" && exit 1",
+        "start": "node --experimental-specifier-resolution=node ./build/index.js",
+        "start:dev": "nodemon ./src/index.ts",
         "lint": "eslint ./src/**"
     },
     "keywords": [],
@@ -28,6 +27,7 @@
     "dependencies": {
         "@types/socket.io": "^3.0.2",
         "@types/websocket": "^1.0.5",
+        "@types/ws": "^8.5.3",
         "jimp": "^0.16.1",
         "robotjs": "^0.6.0",
         "ws": "^8.8.0"
