+++ tsconfig.json
@@ -2,18 +2,18 @@
   "compilerOptions": {
 
     /* Language and Environment */
-    "target": "ES6",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
+    "target": "ES5",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
     "experimentalDecorators": true,                      /* Enable experimental support for TC39 stage 2 draft decorators. */
     "emitDecoratorMetadata": true,                       /* Emit design-type metadata for decorated declarations in source files. */
 
     /* Modules */
-    "module": "CommonJS",
+    "module": "commonjs",
     /* Specify what module code is generated. */
     "moduleResolution": "node",                          /* Specify how TypeScript looks up a file from a given module specifier. */
 
     /* Emit */
     "sourceMap": true,                                   /* Create source map files for emitted JavaScript files. */
-    "outDir": "./build",                                  /* Specify an output folder for all emitted files. */
+    "outDir": "./build/",                                  /* Specify an output folder for all emitted files. */
     "removeComments": true,                              /* Disable emitting comments. */
     "esModuleInterop": true,                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
     "forceConsistentCasingInFileNames": true,            /* Ensure that casing is correct in imports. */
@@ -21,6 +21,11 @@
     /* Type Checking */
     "strict": true,                                      /* Enable all strict type-checking options. */
     "strictPropertyInitialization": false,               /* Check for class properties that are declared but not set in the constructor. */
-    "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
-  }
+    "skipLibCheck": true,                                 /* Skip type checking all .d.ts files. */
+
+  },
+  "include": ["./src/*.ts"],
+  "exclude": [
+    "node_modules"
+  ]
 }
